requirements.txt
geopandas>=0.14.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
contextily>=1.5.0
